Sniffing/scripts/requirements.txt
scapy
python-dateutil
pycryptodome
locust